python-app/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pdf2image==1.16.3
pydantic==2.4.2
pillow==11.1.0
pdfplumber==0.10.4
PyMuPDF==1.25.2

# PyMuPDF replacement

# Add comment that poppler-utils is installed via apt-get
# The Python package wasn't needed
# poppler-utils>=0.1.0

# Flask is not needed if you're using FastAPI
# Flask==2.3.3